requirements.txt
email_validator==2.1.1
Flask==3.0.3
flask_login==0.6.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
